move locale switcher to client folder
