+++ app/[locale]/(custom)/layout.tsx
@@ -1,6 +1,6 @@
-import { Box, Flex, HStack } from '@/components/chakra';
-import { Navbar } from '@/components/client';
-import { LocaleSwitcher, ProfileMenu } from '@/components/server';
+import { LocaleSwitcher, Navbar } from '@/components/client';
+import { Box, Flex } from '@/components/chakra';
+import { ProfileMenu } from '@/components/server';
 import type { ReactNode } from '@/types';
 import type { Locale } from '@/configs/i18n.config';
 
@@ -14,14 +14,10 @@ export default function CustomLayout({
   return (
     <Flex direction="column">
       <Navbar>
-        <HStack spacing={1}>
-          <Box>
-            <LocaleSwitcher locale={params.locale} />
-          </Box>
-          <Box>
-            <ProfileMenu />
-          </Box>
-        </HStack>
+        <Flex gap={1}>
+          <LocaleSwitcher locale={params.locale} />
+          <ProfileMenu />
+        </Flex>
       </Navbar>
 
       <Box padding={6} width="100%" height="100vh" overflowY="scroll">
