+++ components/server/index.ts
@@ -1,2 +1,2 @@
-export * from './ProfileMenu';
 export * from './CustomHead';
+export * from './ProfileMenu';
